+++ iidm/iidm-modification/src/main/java/com/powsybl/iidm/modification/topology/CreateLineOnLineBuilder.java
@@ -19,7 +19,7 @@ public class CreateLineOnLineBuilder {
     private Line line = null;
     private LineAdder lineAdder = null;
 
-    private double percent = 50;
+    private double positionPercent = 50;
 
     private String fictitiousVlId = null;
     private String fictitiousVlName = null;
@@ -34,13 +34,13 @@ public class CreateLineOnLineBuilder {
     private String line2Name = null;
 
     public CreateLineOnLine build() {
-        return new CreateLineOnLine(percent, bbsOrBusId, fictitiousVlId, fictitiousVlName,
+        return new CreateLineOnLine(positionPercent, bbsOrBusId, fictitiousVlId, fictitiousVlName,
                 createFictSubstation, fictitiousSubstationId, fictitiousSubstationName,
                 line1Id, line1Name, line2Id, line2Name, line, lineAdder);
     }
 
-    public CreateLineOnLineBuilder withPercent(double percent) {
-        this.percent = percent;
+    public CreateLineOnLineBuilder withPositionPercent(double positionPercent) {
+        this.positionPercent = positionPercent;
         return this;
     }
